+++ jannovar-filter/src/test/java/de/charite/compbio/jannovar/filter/impl/facade/TresholdFilterTestBase.java
@@ -1,0 +1,15 @@
+package de.charite.compbio.jannovar.filter.impl.facade;
+
+import de.charite.compbio.jannovar.filter.impl.gt.GenotypeFilterTestBase;
+
+/**
+ * Intermediate class, we can take the implementation from GenotypeFilterTestBase
+ * 
+ * Sub classing this class for semantic reasons so the threshold filter tests don't have to extend the genotype filter
+ * test base directly
+ *
+ * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
+ */
+public class TresholdFilterTestBase extends GenotypeFilterTestBase {
+
+}