+++ jannovar-vardbs/src/test/java/de/charite/compbio/jannovar/vardbs/clinvar/ClinVarVCFHeaderExtenderTest.java
@@ -9,10 +9,6 @@ import htsjdk.variant.vcf.VCFHeader;
 
 public class ClinVarVCFHeaderExtenderTest {
 
-	@Before
-	public void setUp() throws Exception {
-	}
-
 	@Test
 	public void test() {
 		VCFHeader header = new VCFHeader();